ノルマのUIをスライダーにした。次のWaveになったらタイマーリセット

+++ Assets/Scripts/UI/ShowScoreModel.cs
@@ -7,16 +7,16 @@ public class ShowScoreModel
     public IObservable<int> CurrentScoreChanged => _score;
     readonly ReactiveProperty<int> _score = new();
 
-    public int Count { get => _count.Value; set => _count.Value = value; }
-    public IObservable<int> CurrentCountChanged => _count;
-    readonly ReactiveProperty<int> _count = new();
+    public float Count { get => _count.Value; set => _count.Value = value; }
+    public IObservable<float> CurrentCountChanged => _count;
+    readonly ReactiveProperty<float> _count = new();
 
     public void SetScoreValue(int value)
     {
         _score.Value = value;
     }
 
-    public void SetCountValue(int value)
+    public void SetCountValue(float value)
     {
         _count.Value = value;
     }
